Fix wrong provider registry updates.
This is a multi-device provider, so adding a device does
not automatically mean that the provider gets enabled (there
might be an existing device), nor that removing a device disables
the provider (there might be remaining devices).

This adds an additional check to count the number of devices
*after* the state change to only react when the first device
has been added and when the last one is removed.

Signed-off-by: Christoph Wurst <christoph@winzerhof-wurst.at>